chore: split worker pool and task queue implementation
