+++ src/task_queue/array_queue.ts
@@ -1,0 +1,25 @@
+import assert from 'node:assert';
+
+import type { TaskQueue, Task } from '.';
+
+export class ArrayTaskQueue implements TaskQueue {
+  tasks: Task[] = []
+
+  get size () {
+    return this.tasks.length;
+  }
+
+  shift (): Task | null {
+    return this.tasks.shift() as Task;
+  }
+
+  push (task: Task): void {
+    this.tasks.push(task);
+  }
+
+  remove (task: Task): void {
+    const index = this.tasks.indexOf(task);
+    assert.notStrictEqual(index, -1);
+    this.tasks.splice(index, 1);
+  }
+}
